pyproject.toml
[project]
name = "llmchatassistant"
version = "0.1.0"
description = "智能网络运维助手 - 基于 NPLT 协议的 AI 对话和 RDT 可靠文件传输系统"
requires-python = ">=3.11"
dependencies = [
    "zai-sdk>=0.1.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "sniffio>=1.3.0",
    "faiss-cpu>=1.7.0",
    "aiohttp>=3.9.0",
    "prompt-toolkit>=3.0.0",  # 修复中文输入法全半角问题
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clients", "server", "shared"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "psutil>=5.9.0",
]
